+++ Editor/Scripts/ObjectPoolEditor.cs
@@ -9,17 +9,11 @@ namespace FredericRP.ObjectPooling
   {
     public override void OnInspectorGUI()
     {
-      ObjectPool pool = target as ObjectPool;
-
       // New : set ID of ObjectPool
-      //EditorGUILayout.BeginHorizontal(EditorStyles.toolbar);
       EditorGUILayout.PropertyField(serializedObject.FindProperty("id"));
-      //EditorGUILayout.EndHorizontal();
       //
-      List<ObjectPool.PoolGameObjectInfo> poolObjectList = pool.PoolGameObjectInfoList;
+      SerializedProperty poolObjectListProperty = serializedObject.FindProperty("poolObjectList");
       Color previousColor = GUI.color;
-      //*
-      //Rect defaultRect = EditorGUILayout.GetControlRect(false, GUILayout.ExpandWidth(true));
 
       // EditorGUILayout.GetControlRect adds a new Layout in the grid and prevent using its width simpluy so we just use the total view and substract usual extra space
       float totalWidth = EditorGUIUtility.currentViewWidth - 24;
@@ -31,83 +25,77 @@ namespace FredericRP.ObjectPooling
       float buttonWidth = 20;
       float buttonsWidth = 4 * buttonWidth;
       // min max slider use an int field of size 38, for id, parent and slider we are left with: (total width - 2x38) / 3
-      idWidth = (totalWidth - 2 * minMaxIntWidth - prefabWidth - buttonsWidth - 9 * EditorGUIUtility.standardVerticalSpacing) / 3;
+      idWidth = (totalWidth - 2 * minMaxIntWidth - prefabWidth - buttonsWidth) / 3 - 3 * EditorGUIUtility.standardVerticalSpacing;
       parentWidth = idWidth;
       sliderWidth = idWidth;
 
       EditorGUILayout.BeginHorizontal();
-      GUILayout.Label("object id", EditorStyles.boldLabel, GUILayout.Width(idWidth));
-      GUILayout.Label("prefab", EditorStyles.boldLabel, GUILayout.Width(prefabWidth));
-      GUILayout.Label("default", EditorStyles.boldLabel, GUILayout.Width(minMaxIntWidth + sliderWidth / 2));
-      GUIStyle rightLabel = new GUIStyle(EditorStyles.boldLabel);
+      GUIStyle headerStyle = EditorStyles.miniBoldLabel;
+      GUILayout.Label("object id", headerStyle, GUILayout.Width(idWidth));
+      GUILayout.Label("prefab", headerStyle, GUILayout.Width(prefabWidth));
+      GUILayout.Label("default", headerStyle, GUILayout.Width(minMaxIntWidth + sliderWidth / 2 + EditorGUIUtility.standardVerticalSpacing));
+      GUIStyle rightLabel = new GUIStyle(headerStyle);
       rightLabel.alignment = TextAnchor.MiddleRight;
       GUILayout.Label("max count", rightLabel, GUILayout.Width(minMaxIntWidth + sliderWidth / 2));
-      GUILayout.Label("def. parent", EditorStyles.boldLabel, GUILayout.Width(parentWidth));
-      //GUILayout.Label("- - - -", GUILayout.Width(buttonsWidth));
+      GUILayout.Label("def. parent", headerStyle, GUILayout.Width(parentWidth));
       EditorGUILayout.EndHorizontal();
-      //EditorGUI.DrawRect(new Rect(0, 0, idWidth, EditorGUIUtility.singleLineHeight), Color.green);
-      //EditorGUI.DrawRect(new Rect(18, 20, totalWidth, EditorGUIUtility.singleLineHeight), Color.cyan);
-      //EditorGUI.DrawRect(new Rect(18 + EditorGUIUtility.standardVerticalSpacing + idWidth, EditorGUIUtility.singleLineHeight + 4, prefabWidth, EditorGUIUtility.singleLineHeight), Color.yellow);
 
       EditorGUI.BeginChangeCheck();
 
-      if (poolObjectList.Count > 0)
+      if (poolObjectListProperty.isArray && poolObjectListProperty.arraySize > 0)
       {
-        for (int i = 0; i < poolObjectList.Count; i++)
+        for (int i = 0; i < poolObjectListProperty.arraySize; i++)
         {
-          ObjectPool.PoolGameObjectInfo data = poolObjectList[i];
+          SerializedProperty dataProperty = poolObjectListProperty.GetArrayElementAtIndex(i);
           EditorGUILayout.BeginHorizontal();
 
           // PoolObject : name, prefab, bufferCount, defaultParent
-          data.id = EditorGUILayout.TextField(data.id, GUILayout.Width(idWidth));
-          //data.tag = EditorGUILayout.TagField(data.tag);
-          if (data.prefab == null)
+          EditorGUILayout.PropertyField(dataProperty.FindPropertyRelative("id"), GUIContent.none, true, GUILayout.Width(idWidth));
+          SerializedProperty prefabProperty = dataProperty.FindPropertyRelative("prefab");
+          if (prefabProperty.objectReferenceValue == null)
             GUI.color = Color.red;
-          data.prefab = (GameObject)EditorGUILayout.ObjectField(data.prefab, typeof(GameObject), false, GUILayout.Width(prefabWidth));
+          EditorGUILayout.PropertyField(prefabProperty, GUIContent.none, true, GUILayout.Width(prefabWidth));
           GUI.color = previousColor;
 
-          data.bufferCount = EditorGUILayout.IntField(data.bufferCount, GUILayout.Width(minMaxIntWidth));
-          float count = data.bufferCount;
-          float max = data.maxCount;
+          SerializedProperty bufferCountProperty = dataProperty.FindPropertyRelative("bufferCount");
+          SerializedProperty maxCountProperty = dataProperty.FindPropertyRelative("maxCount");
+          EditorGUILayout.PropertyField(bufferCountProperty, GUIContent.none, true, GUILayout.Width(minMaxIntWidth));
+          float count = bufferCountProperty.intValue;
+          float max = maxCountProperty.intValue;
           EditorGUILayout.MinMaxSlider(ref count, ref max, 0, ObjectPool.MAX_BUFFER, GUILayout.Width(sliderWidth));
-          data.bufferCount = (int)count;
-          data.maxCount = (int)max;
-          data.maxCount = EditorGUILayout.IntField(data.maxCount, GUILayout.Width(minMaxIntWidth));
+          bufferCountProperty.intValue = (int)count;
+          maxCountProperty.intValue = (int)max;
+          EditorGUILayout.PropertyField(maxCountProperty, GUIContent.none, true, GUILayout.Width(minMaxIntWidth));
 
-          //GUILayout.Label("on");
-          if (data.defaultParent == null)
+          SerializedProperty defaultParentProperty = dataProperty.FindPropertyRelative("defaultParent");
+          if (defaultParentProperty.objectReferenceValue == null)
             GUI.color = Color.red;
-          data.defaultParent = (Transform)EditorGUILayout.ObjectField(data.defaultParent, typeof(Transform), true, GUILayout.Width(parentWidth));
+          EditorGUILayout.PropertyField(defaultParentProperty, GUIContent.none, true, GUILayout.Width(parentWidth));
           GUI.color = previousColor;
 
           GUI.enabled = (i > 0);
           if (GUILayout.Button("\u25B2", EditorStyles.miniButtonLeft, GUILayout.Width(buttonWidth)))
           {
             // Switch with previous property
-            ObjectPool.PoolGameObjectInfo pgoInfo = poolObjectList[i];
-            poolObjectList.RemoveAt(i);
-            poolObjectList.Insert(i - 1, pgoInfo);
-
+            poolObjectListProperty.MoveArrayElement(i, i - 1);
           }
-          GUI.enabled = (i < poolObjectList.Count - 1);
+          GUI.enabled = (i < poolObjectListProperty.arraySize - 1);
           if (GUILayout.Button("\u25BC", EditorStyles.miniButtonMid, GUILayout.Width(buttonWidth)))
           {
             // Switch with next property
-            ObjectPool.PoolGameObjectInfo pgoInfo = poolObjectList[i];
-            poolObjectList.RemoveAt(i);
-            poolObjectList.Insert(i + 1, pgoInfo);
+            poolObjectListProperty.MoveArrayElement(i, i + 1);
           }
           GUI.enabled = true;
           GUI.color = Color.green;
           if (GUILayout.Button("+", EditorStyles.miniButtonMid, GUILayout.Width(buttonWidth)))
           {
-            poolObjectList.Insert(i + 1, new ObjectPool.PoolGameObjectInfo());
+            poolObjectListProperty.InsertArrayElementAtIndex(i + 1);
           }
           GUI.color = Color.red;
           if (GUILayout.Button("X", EditorStyles.miniButtonRight, GUILayout.Width(buttonWidth)))
           {
             // Remove property
-            poolObjectList.RemoveAt(i);
+            poolObjectListProperty.DeleteArrayElementAtIndex(i);
           }
           GUI.color = previousColor;
           EditorGUILayout.EndHorizontal();
@@ -119,14 +107,15 @@ namespace FredericRP.ObjectPooling
         GUI.color = Color.green;
         if (GUILayout.Button("+", EditorStyles.miniButton, GUILayout.Width(buttonWidth)))
         {
-          poolObjectList.Add(new ObjectPool.PoolGameObjectInfo());
+          poolObjectListProperty.InsertArrayElementAtIndex(0);
+          //poolObjectList.Add(new ObjectPool.PoolGameObjectInfo());
         }
         GUI.color = previousColor;
       }
 
       if (EditorGUI.EndChangeCheck())
       {
-        EditorUtility.SetDirty(pool);
+        serializedObject.ApplyModifiedProperties();
       }
     }
   }